api/requirements.txt
alembic==1.2.1
ansi2html==1.5.2
asn1crypto==1.0.1
autopep8==1.4
backports-abc==0.5
boto3==1.9.229
botocore==1.12.230
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.4.2
dnspython==1.16.0
docutils==0.14
email-validator==1.0.3
enum34==1.1.6
expiringdict==1.1.4
functools32==3.2.3.post2
futures==3.2.0
idna==2.8
ipaddress==1.0.22
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.16.3
psycopg2==2.7.5
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pystache==0.5.4
python-dateutil==2.7.3
python-editor==1.0.4
PyYAML==3.13
requests==2.22.0
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.11
stripe==2.28.2
tornado==5.1
unicodecsv==0.14.1
urllib3==1.23
pytest==4.6.9
hypothesis==4.57.1
pytest-tornado==0.8.0
coverage==5.0.3
python-jsonschema-objects==0.3.12
pytest-mock==2.0.0
pinject==0.14.1
